Fixing errors I saw when it was trying to clean up files, but not applicable when there are no local files and it's only a manifest.

+++ stash_engine/lib/stash/repo/repository.rb
@@ -129,6 +129,7 @@ module Stash
       end
 
       def remove_if_exists(file)
+        return if file.blank?
         FileUtils.remove_entry_secure(file) if File.exist?(file)
       end
